requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.5
django-ckeditor==6.7.2
django-cors-headers==4.3.1
django-environ==0.12.0
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
email_validator==2.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
googleapis-common-protos==1.66.0
googletrans==4.0.2
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2023.4
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
whitenoise==6.8.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.5
django-ckeditor==6.7.2
django-cors-headers==4.3.1
django-environ==0.12.0
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
email_validator==2.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
googleapis-common-protos==1.66.0
googletrans==4.0.2
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2023.4
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
whitenoise==6.8.2

